updated the badges verification with utility and constant file
